+++ src/pdfnaut/objects/catalog.py
@@ -163,7 +163,7 @@ class ViewerPreferences(PdfDictionary):
     If this value is none, the document producer may choose their own default setting.
     """
 
-    print_page_range: PdfArray[int] | None = field(default=None)
+    print_page_range: list[int] | None = field(default=None, encoder=PdfArray, decoder=list)
     """The page numbers used to initialize the print dialog box. The array should
     contain an even number of values interpreted as pairs, with each pair specifying
     the first and last pages in a sub-range of pages to be printed (the first page being